requirements.txt
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
cssbeautifier==1.14.7
Django==4.1.7
django-crispy-forms==2.0
django-formtools==2.4
django-otp==1.1.6
django-phonenumber-field==6.4.0
django-two-factor-auth==1.15.1
djlint==1.19.16
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.14.7
pathspec==0.11.0
phonenumbers==8.13.7
pypng==0.20220715.0
PyYAML==6.0
qrcode==7.4.2
regex==2022.10.31
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.7
